requirements.txt
Django>=5.2.0
djangorestframework>=3.15.0
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
django-filter>=23.5.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
